# docs/swagger.yaml
definitions:
...
  models.ChatFeedbackRequest:
    properties:
      feedback:
        type: string
    required:
    - feedback
    type: object
  models.ChatMessage:
    properties:
      feedback:
        type: string
      message:
        type: string
      message_id:
        type: integer
      rating:
        type: integer
      sender:
        type: string
      time_sent:
        type: string
      user_id:
        type: integer
    type: object
  models.ChatMessageRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  models.ChatRatingRequest:
    properties:
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
...
  utils.HTTPError:
    properties:
      code:
        type: integer
      error:
        type: string
      title:
        type: string
    type: object
...
host: https://user-api-production-99c2.up.railway.app/
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User API
  version: "1.0"
paths:
...
  /chat:
    get:
      consumes:
      - application/json
      description: Gets The messages of the last 2 days
      produces:
      - text/plain
      responses:
        "200":
          description: messages
          schema:
            items:
              $ref: '#/definitions/models.ChatMessage'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Get recent messages
      tags:
      - Chat
    post:
      consumes:
      - application/json
      description: Send a new message to the assitant
      parameters:
      - description: User Registration Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChatMessageRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Ai assitant response
          schema:
            $ref: '#/definitions/models.ChatMessage'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Send a new message to the assitant
      tags:
      - Chat
  /chat/{message_id}/feedback:
    put:
      consumes:
      - application/json
      description: Add feedback to a message so the ai can better generate a response
      parameters:
      - description: Rating
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChatFeedbackRequest'
      - description: Message Id
        in: path
        name: message_id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Add feedback to a message
      tags:
      - Chat
  /chat/{message_id}/rate:
    put:
      consumes:
      - application/json
      description: Add rating to a message beteween 1 and 5
      parameters:
      - description: Rating
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChatRatingRequest'
      - description: Message Id
        in: path
        name: message_id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Add rating to a message
      tags:
      - Chat
...
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"